add support for the pgx stdlib driver
Add support for the pgx Postgres driver. This only supports the stdlib
driver, which integrates with the standard Go sql package. Usage of the
custom pgx driver is still not supported.

As part of supporting pgx, copyist needed to support the QueryContext,
ExecContext, and BeginTx driver functions. The tests are also refactored
in order to share common driver-specific tests.
