+++ drivertest/commontest/testdata/common_test.copyist
@@ -1,0 +1,17 @@
+1=DriverOpen	1:nil
+2=ConnPrepare	2:"SHOW session_id"	1:nil
+3=StmtNumInput	3:0
+4=StmtQuery	1:nil
+5=RowsColumns	9:["session_id"]
+6=RowsNext	11:[2:"1662ea276e9e47a80000000000000001"]	1:nil
+7=RowsNext	11:[]	7:EOF
+8=ConnPrepare	2:"SELECT name FROM customers WHERE id=$1"	1:nil
+9=StmtNumInput	3:1
+10=RowsColumns	9:["name"]
+11=RowsNext	11:[2:"Andy"]	1:nil
+12=RowsNext	11:[2:"1662ea27790bd7dc0000000000000001"]	1:nil
+
+"TestIsOpen"=
+"TestPooling/ensure_connections_are_pooled_within_same_copyist_session"=1,2,3,4,5,6,7,2,3,4,5,6,7
+"TestIndirectOpen"=1,8,9,4,10,11
+"TestPooling/ensure_connections_are_*not*_pooled_across_copyist_sessions"=1,2,3,4,5,12
