+++ drivertest/pgxtest/testdata/pgxtest_test.copyist
@@ -1,0 +1,37 @@
+1=DriverOpen	1:nil
+2=ConnPrepare	2:"SELECT 1::float, 1.1::float, 1e20::float"	1:nil
+3=StmtNumInput	3:0
+4=StmtQuery	1:nil
+5=RowsColumns	9:["float8","float8","float8"]
+6=RowsNext	11:[5:1,5:1.1,5:1e+20]	1:nil
+7=ConnPrepare	2:"\n\t\tCREATE TABLE datatypes\n\t\t(i INT, s TEXT, tz TIMESTAMPTZ, t TIMESTAMP, b BOOL,\n\t\t by BYTES, f FLOAT, d DECIMAL, fa FLOAT[], u UUID)\n\t"	1:nil
+8=StmtExec	1:nil
+9=ConnPrepare	2:"\n\t\tINSERT INTO datatypes VALUES\n\t\t\t(1, 'foo' || CHR(9) || CHR(10) || ' ,]', '2000-01-01T10:00:00Z', '2000-01-01T10:00:00Z',\n\t\t\t true, 'ABCD', 1.1, 100.1234, ARRAY(1.1, 1.2345678901234567),\n\t\t\t '8B78978B-7D8B-489E-8CA9-AC4BDC495A82'),\n\t\t\t(2, '', '2000-02-02T11:11:11-08:00', '2000-02-02T11:11:11-08:00', false,\n\t\t\t '', -1e10, -0.0, ARRAY(), '00000000-0000-0000-0000-000000000000')\n\t"	1:nil
+10=ResultRowsAffected	4:0	1:nil
+11=ConnPrepare	2:"SELECT i, s, tz, t, b, by, f, d, fa, u FROM datatypes"	1:nil
+12=RowsColumns	9:["i","s","tz","t","b","by","f","d","fa","u"]
+13=RowsNext	11:[4:1,2:"foo\t\n ,]",8:2000-01-01T02:00:00-08:00,8:2000-01-01T10:00:00Z,6:true,10:QUJDRA,5:1.1,2:"100.1234",2:"{1.1,1.2345678901234567}",2:"8b78978b-7d8b-489e-8ca9-ac4bdc495a82"]	1:nil
+14=RowsNext	11:[4:2,2:"",8:2000-02-02T11:11:11-08:00,8:2000-02-02T11:11:11Z,6:false,10:,5:-1e+10,2:"0.0",2:"{}",2:"00000000-0000-0000-0000-000000000000"]	1:nil
+15=ConnBegin	1:nil
+16=ConnPrepare	2:"INSERT INTO customers VALUES ($1, $2)"	1:nil
+17=StmtNumInput	3:2
+18=TxCommit	1:nil
+19=TxRollback	1:nil
+20=ConnPrepare	2:"SELECT COUNT(*) FROM customers"	1:nil
+21=RowsColumns	9:["count"]
+22=RowsNext	11:[4:4]	1:nil
+23=RowsNext	11:[]	7:EOF
+24=ConnPrepare	2:"SELECT name FROM customers WHERE id=$1"	1:nil
+25=StmtNumInput	3:1
+26=RowsColumns	9:["name"]
+27=RowsNext	11:[2:"Andy"]	1:nil
+28=ResultRowsAffected	4:1	1:nil
+29=ConnPrepare	2:"SELECT name FROM customers WHERE id=?"	1:nil
+
+"TestTxns"=1,15,16,17,8,18,15,16,17,8,19,20,3,4,21,22,23
+"TestQuery"=1,24,25,4,26,27,23
+"TestInsert"=1,16,17,8,28,20,3,4,21,22,23
+"TestFloatLiterals/run_1"=1,2,3,4,5,6
+"TestSqlx"=1,15,29,25,4,26,27,23,18
+"TestFloatLiterals/run_2"=1,2,3,4,5,6
+"TestDataTypes"=1,7,3,8,9,3,8,10,11,3,4,12,13,14
