+++ go.mod
@@ -3,9 +3,6 @@ module github.com/cockroachdb/copyist
 go 1.15
 
 require (
-	github.com/fortytw2/leaktest v1.3.0
-	github.com/jmoiron/sqlx v1.2.0
-	github.com/lib/pq v1.7.0
+	github.com/jmoiron/sqlx v1.3.1
 	github.com/stretchr/testify v1.7.0
-	google.golang.org/appengine v1.6.6 // indirect
 )
